ipl/requirement.txt
amqp==2.5.1
billiard==3.5.0.5
celery==4.1.1
Django==2.2.6
django-cors-headers==3.1.1
django-model-utils==3.2.0
djangorestframework==3.10.3
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
numpy==1.17.3
pandas==0.25.3
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.3
raven==6.9.0
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
uuid==1.30
vine==1.3.0
zipp==0.6.0